Added NONE value
Is good to add and none value if we don't have any arrow direction to any bubble layout. In my case I have some bubbles with arrows and some not but when I try with switch I can use ArrowDirection.NONE. If this will added it will be very good.